requirements.txt
bcrypt==3.1.6
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.2.16
SQLAlchemy-Utils==0.33.11
Werkzeug==0.14.1
